Converted quickCom to use resizeBuff()
resizeBuff was developed for quickCom. But it never was updated to the "production" version and just ran the prototype. Until now. Its finally using the resizeBuff library. Yay!

Also pulled  cellCommon.h out. The IDE STILL pulls in unwanted code when compiling if it see this. Even with .a linking activated.